Extract `FormHelpers` extension methods class from `EditContactModel`
This refactoring reduces the responsibility of the `EditContactModel`
class, allowing for it to be more easily made generic over both schools
and trusts.
