+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Contacts/_EditContactsForm.cshtml
@@ -1,4 +1,5 @@
-﻿@model EditContactModel
+﻿@using DfE.FindInformationAcademiesTrusts.Pages.Shared
+@model EditContactModel
 
 <partial name="Trusts/_TrustBanner" model="@Model"/>
